feat(plugins): load plugin description from package
