+++ elasticai/creator/plugins/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["Plugin", "read_plugins_from_package"]
+
+from .plugin import Plugin, read_plugins_from_package
